Reading gps message types
